refactor: unlink tradeWidgetAddons module from bridge
