+++ apps/cowswap-frontend/src/modules/tradeWidgetAddons/containers/TradeRateDetails/index.tsx
@@ -1,11 +1,7 @@
-import { useMemo, useState, useCallback, ReactElement } from 'react'
+import { useMemo, useState, useCallback, ReactElement, ReactNode } from 'react'
 
-import { TokenWithLogo } from '@cowprotocol/common-const'
-import { useIsBridgingEnabled } from '@cowprotocol/common-hooks'
-import { useWalletDetails } from '@cowprotocol/wallet'
 import { CurrencyAmount } from '@uniswap/sdk-core'
 
-import { BridgeData, BridgeRouteBreakdown, BridgeProtocolConfig, BridgeAccordionSummary } from 'modules/bridge'
 import {
   getTotalCosts,
   TradeFeesAndCosts,
@@ -14,8 +10,6 @@ import {
   NetworkCostsRow,
   useReceiveAmountInfo,
   useShouldPayGas,
-  useIsHooksTradeType,
-  useIsCurrentTradeBridging,
 } from 'modules/trade'
 import { useTradeQuote } from 'modules/tradeQuote'
 import { useIsSlippageModified, useTradeSlippage } from 'modules/tradeSlippage'
@@ -35,9 +29,17 @@ interface TradeRateDetailsProps {
   rateInfoParams: RateInfoParams
   children?: ReactElement
   isTradePriceUpdating: boolean
+  accordionContent?: ReactNode
+  feeWrapper?: (feeElement: ReactNode) => ReactNode
 }
 
-export function TradeRateDetails({ rateInfoParams, deadline, isTradePriceUpdating }: TradeRateDetailsProps) {
+export function TradeRateDetails({
+  rateInfoParams,
+  deadline,
+  isTradePriceUpdating,
+  accordionContent,
+  feeWrapper,
+}: TradeRateDetailsProps) {
   const [isFeeDetailsOpen, setFeeDetailsOpen] = useState(false)
 
   const slippage = useTradeSlippage()
@@ -46,22 +48,8 @@ export function TradeRateDetails({ rateInfoParams, deadline, isTradePriceUpdatin
   const derivedTradeState = useDerivedTradeState()
   const tradeQuote = useTradeQuote()
   const shouldPayGas = useShouldPayGas()
-  const isHooksTabEnabled = useIsHooksTradeType()
-  const { isSmartContractWallet } = useWalletDetails()
-
-  // Bridge related state
-  const isBridgingEnabled = useIsBridgingEnabled(isSmartContractWallet)
-  const isCurrentTradeBridging = useIsCurrentTradeBridging()
-  const showBridgeUI = isBridgingEnabled && isCurrentTradeBridging
-  // TODO: bridgeDetailsUI: Set a real value for bridgeData based on bridging logic
-  const bridgeData = null as BridgeData | null
-  const providerDetails: BridgeProtocolConfig | undefined = bridgeData?.bridgeProvider
-  const bridgeEstimatedTime: number | undefined = bridgeData?.estimatedTime
 
   const inputCurrency = derivedTradeState?.inputCurrency
-  const outputCurrency = derivedTradeState?.outputCurrency
-  const inputCurrencyAmount = derivedTradeState?.inputCurrencyAmount
-  const outputCurrencyAmount = derivedTradeState?.outputCurrencyAmount
 
   const costsExceedFeeRaw = tradeQuote.error instanceof QuoteApiError ? tradeQuote?.error?.data?.fee_amount : undefined
 
@@ -95,70 +83,33 @@ export function TradeRateDetails({ rateInfoParams, deadline, isTradePriceUpdatin
 
   const totalCosts = getTotalCosts(receiveAmountInfo)
 
-  const shouldRenderBridgeUI = !!(
-    showBridgeUI &&
-    bridgeData &&
-    !isHooksTabEnabled &&
-    inputCurrency &&
-    outputCurrency &&
-    inputCurrencyAmount &&
-    outputCurrencyAmount &&
-    'sellAmount' in bridgeData &&
-    'buyAmount' in bridgeData
-  )
-
-  const accordionContent = shouldRenderBridgeUI ? (
-    <BridgeRouteBreakdown
-      sellCurrencyAmount={inputCurrencyAmount as CurrencyAmount<TokenWithLogo>}
-      buyCurrencyAmount={outputCurrencyAmount as CurrencyAmount<TokenWithLogo>}
-      bridgeReceiveCurrencyAmount={outputCurrencyAmount as CurrencyAmount<TokenWithLogo>}
-      networkCost={bridgeData.networkCost}
-      swapMinReceive={bridgeData.swapMinReceive}
-      swapExpectedToReceive={bridgeData.swapExpectedToReceive}
-      swapMaxSlippage={bridgeData.swapMaxSlippage}
-      bridgeFee={bridgeData.bridgeFee}
-      maxBridgeSlippage={bridgeData.maxBridgeSlippage}
-      estimatedTime={bridgeData.estimatedTime}
-      recipient={bridgeData.recipient}
-      bridgeProvider={bridgeData.bridgeProvider}
-    />
-  ) : (
-    <>
-      <TradeFeesAndCosts
-        receiveAmountInfo={receiveAmountInfo}
-        withTimelineDot={false}
-        networkCostsSuffix={shouldPayGas ? <NetworkCostsSuffix /> : null}
-        networkCostsTooltipSuffix={<NetworkCostsTooltipSuffix />}
-        volumeFeeTooltip={volumeFeeTooltip}
-      />
-      {slippage && (
-        <RowSlippage
-          isTradePriceUpdating={isTradePriceUpdating}
-          allowedSlippage={slippage}
-          isSlippageModified={isSlippageModified}
-        />
-      )}
-      <RowDeadline deadline={deadline} />
-    </>
-  )
-
   return (
     <TradeTotalCostsDetails
       totalCosts={totalCosts}
       rateInfoParams={rateInfoParams}
       isFeeDetailsOpen={isFeeDetailsOpen}
       toggleAccordion={toggleAccordion}
-      feeWrapper={
-        shouldRenderBridgeUI
-          ? (feeElement) => (
-              <BridgeAccordionSummary bridgeEstimatedTime={bridgeEstimatedTime} bridgeProtocol={providerDetails}>
-                {feeElement}
-              </BridgeAccordionSummary>
-            )
-          : undefined
-      }
+      feeWrapper={feeWrapper}
     >
-      {accordionContent}
+      {accordionContent || (
+        <>
+          <TradeFeesAndCosts
+            receiveAmountInfo={receiveAmountInfo}
+            withTimelineDot={false}
+            networkCostsSuffix={shouldPayGas ? <NetworkCostsSuffix /> : null}
+            networkCostsTooltipSuffix={<NetworkCostsTooltipSuffix />}
+            volumeFeeTooltip={volumeFeeTooltip}
+          />
+          {slippage && (
+            <RowSlippage
+              isTradePriceUpdating={isTradePriceUpdating}
+              allowedSlippage={slippage}
+              isSlippageModified={isSlippageModified}
+            />
+          )}
+          <RowDeadline deadline={deadline} />
+        </>
+      )}
     </TradeTotalCostsDetails>
   )
 }
